Integrate symmetric cipher demo into main game logic
Symmetric cipher that shows students how various operations can be used to create your own symmetric encryption, and how these encryptions are reversed to get plaintext. In this case, the key is the exact sequence of operations used to encode plaintext. This provides visual demonstration on how the same operations can be used to encode and decode plaintext.